Separated Boss login from the mail file into shrines and boss manager class
